Types.h : TFrequency : becomes of type "u32" / "Hz" (instead of "float" / "MHz" ) CWifi : Attenuation : log10(A/1000) = log10(A)-log10(1000) = log10(A)-3

+++ src/types.h
@@ -43,8 +43,8 @@ const TPower TPower_MIN=INT8_MIN;
 typedef float TDistance; // in meters
 typedef float TScale;
 
-// float
-typedef float TFrequency;
+// u32
+typedef u32 TFrequency; // Hz
 
 // unsigned short
 typedef u16 TMinimalSize;
